requirements.txt
python-dotenv==1.0.1
aiogram==3.6.0
beautifulsoup4==4.12.3
requests==2.31.0
SQLAlchemy==2.0.30
openpyxl==3.1.2